Check and fix support of ClearPass < 6.8.0

+++ Tests/integration/StaticHostList.Tests.ps1
@@ -8,55 +8,63 @@
 Describe  "Get Static Host List" {
 
     BeforeAll {
-        #Add 2 entries
-        Add-ArubaCPStaticHostList -name pester_SHL-list-IPAddress -host_format list -host_type IPAddress -host_entries_address 192.0.2.1 -host_entries_description "Add via PowerArubaCP"
-        Add-ArubaCPStaticHostList -name pester_SHL-list-MACAddress -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05 -host_entries_description "Add via PowerArubaCP"
+        if ($VersionBefore680 -eq 0) {
+            #Add 2 entries
+            Add-ArubaCPStaticHostList -name pester_SHL-list-IPAddress -host_format list -host_type IPAddress -host_entries_address 192.0.2.1 -host_entries_description "Add via PowerArubaCP"
+            Add-ArubaCPStaticHostList -name pester_SHL-list-MACAddress -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05 -host_entries_description "Add via PowerArubaCP"
+        }
     }
 
-    It "Get Static Host List Does not throw an error" {
+    It "Get Static Host List Does not throw an error" -Skip:$VersionBefore680 {
         {
             Get-ArubaCPStaticHostList
         } | Should Not Throw
     }
 
-    It "Get ALL Static Host List" {
+    It "Get ALL Static Host List" -Skip:$VersionBefore680 {
         $shl = Get-ArubaCPStaticHostList
         $shl.count | Should not be $NULL
     }
 
-    It "Get Static Host List (pester_SHL-list-IPAddress)" {
+    It "Get Static Host List (pester_SHL-list-IPAddress)" -Skip:$VersionBefore680 {
         $shl = Get-ArubaCPStaticHostList | Where-Object { $_.name -eq "pester_SHL-list-IPAddress" }
         $shl.id | Should not be BeNullOrEmpty
         $shl.name | Should be "pester_SHL-list-IPAddress"
     }
 
-    It "Get Static Host List (pester_SHL-list-IPAddress) and confirm (via Confirm-ArubaCPStaticHostList)" {
+    It "Get Static Host List (pester_SHL-list-IPAddress) and confirm (via Confirm-ArubaCPStaticHostList)" -Skip:$VersionBefore680 {
         $shl = Get-ArubaCPStaticHostList | Where-Object { $_.name -eq "pester_SHL-list-IPAddress" }
         Confirm-ArubaCPStaticHostList $shl | Should be $true
     }
 
-    It "Search Static Host List by name (pester_SHL-list-MACAddress)" {
+    It "Search Static Host List by name (pester_SHL-list-MACAddress)" -Skip:$VersionBefore680 {
         $shl = Get-ArubaCPStaticHostList -name pester_SHL-list-MACAddress
         @($shl).count | Should be 1
         $shl.id | Should not be BeNullOrEmpty
         $shl.name | Should be "pester_SHL-list-MACAddress"
     }
 
-    It "Search Static Host List by name (contains *pester*)" {
+    It "Search Static Host List by name (contains *pester*)" -Skip:$VersionBefore680 {
         $shl = Get-ArubaCPStaticHostList -name pester -filter_type contains
         @($shl).count | Should be 2
     }
 
+    It "Get Static Host List License throw a error when use with CPPM <= 6.8.0" -Skip: ($VersionBefore680 -eq 0) {
+        { Get-ArubaCPStaticHostList } | Should throw "Need ClearPass >= 6.8.0 for use this cmdlet"
+    }
+
     AfterAll {
-        #Remove 2 entries
-        Get-ArubaCPStaticHostList -name pester_SHL-list-IPAddress | Remove-ArubaCPStaticHostList -noconfirm
-        Get-ArubaCPStaticHostList -name pester_SHL-list-MACAddress | Remove-ArubaCPStaticHostList -noconfirm
+        if ($VersionBefore680 -eq 0) {
+            #Remove 2 entries
+            Get-ArubaCPStaticHostList -name pester_SHL-list-IPAddress | Remove-ArubaCPStaticHostList -noconfirm
+            Get-ArubaCPStaticHostList -name pester_SHL-list-MACAddress | Remove-ArubaCPStaticHostList -noconfirm
+        }
     }
 }
 
 Describe  "Add Static Host List" {
 
-    It "Add Static Host List with format list and type IPAddress" {
+    It "Add Static Host List with format list and type IPAddress" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1 -host_entries_description "Add via PowerArubaCP"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
         $shl.id | Should not be BeNullOrEmpty
@@ -67,7 +75,7 @@ Describe  "Add Static Host List" {
         $shl.host_entries[0].host_address_desc | Should be "Add via PowerArubaCP"
     }
 
-    It "Add Static Host List with format list and type IPAddress (and add second entries)" {
+    It "Add Static Host List with format list and type IPAddress (and add second entries)" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1 -host_entries_description "Add via PowerArubaCP"
         Get-ArubaCPStaticHostList -name pester_SHL | Add-ArubaCPStaticHostListMember -host_entries_address 192.0.2.2 -host_entries_description "Add via ArubaCPStaticHostListMember"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
@@ -82,7 +90,7 @@ Describe  "Add Static Host List" {
         $shl.host_entries[1].host_address_desc | Should be "Add via ArubaCPStaticHostListMember"
     }
 
-    It "Add Static Host List with format list and type MACAddress" {
+    It "Add Static Host List with format list and type MACAddress" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05 -host_entries_description "Add via PowerArubaCP"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
         $shl.id | Should not be BeNullOrEmpty
@@ -93,7 +101,7 @@ Describe  "Add Static Host List" {
         $shl.host_entries[0].host_address_desc | Should be "Add via PowerArubaCP"
     }
 
-    It "Add Static Host List with format list and type MACAddress (and add second entries)" {
+    It "Add Static Host List with format list and type MACAddress (and add second entries)" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05 -host_entries_description "Add via PowerArubaCP"
         Get-ArubaCPStaticHostList -name pester_SHL | Add-ArubaCPStaticHostListMember -host_entries_address 00:01:02:03:04:06 -host_entries_description "Add via ArubaCPStaticHostListMember"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
@@ -108,7 +116,7 @@ Describe  "Add Static Host List" {
         $shl.host_entries[1].host_address_desc | Should be "Add via ArubaCPStaticHostListMember"
     }
 
-    It "Add Static Host List with format list and type IPAddress (add multiple entries on same time)" {
+    It "Add Static Host List with format list and type IPAddress (add multiple entries on same time)" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1, 192.0.2.2 -host_entries_description "pester entry 1", "pester entry 2"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
         $shl.id | Should not be BeNullOrEmpty
@@ -121,7 +129,7 @@ Describe  "Add Static Host List" {
         $shl.host_entries[1].host_address_desc | Should be "pester entry 2"
     }
 
-    It "Add Static Host List with format list and type MACAddress (add multiple entries on same time without second description)" {
+    It "Add Static Host List with format list and type MACAddress (add multiple entries on same time without second description)" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05, 00:01:02:03:04:06 -host_entries_description "pester entry 1"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
         $shl.id | Should not be BeNullOrEmpty
@@ -134,7 +142,7 @@ Describe  "Add Static Host List" {
         $shl.host_entries[1].host_address_desc | Should be "Add via PowerArubaCP"
     }
 
-    It "Add Static Host List with format list and type IPAddress (and add multiple second entries)" {
+    It "Add Static Host List with format list and type IPAddress (and add multiple second entries)" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1 -host_entries_description "pester entry 1"
         Get-ArubaCPStaticHostList -name pester_SHL | Add-ArubaCPStaticHostListMember -host_entries_address 192.0.2.2, 192.0.2.3 -host_entries_description "Add via ArubaCPStaticHostListMember", "pester entry 3"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
@@ -151,7 +159,7 @@ Describe  "Add Static Host List" {
         $shl.host_entries[2].host_address_desc | Should be "pester entry 3"
     }
 
-    It "Add Static Host List with format list and type MACAddress (and add second entries (without description))" {
+    It "Add Static Host List with format list and type MACAddress (and add second entries (without description))" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05 -host_entries_description "pester entry 1"
         Get-ArubaCPStaticHostList -name pester_SHL | Add-ArubaCPStaticHostListMember -host_entries_address 00:01:02:03:04:06, 00:01:02:03:04:07 -host_entries_description "Add via ArubaCPStaticHostListMember"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
@@ -169,24 +177,28 @@ Describe  "Add Static Host List" {
     }
 
     AfterEach {
-        Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
+        if ($VersionBefore680 -eq 0) {
+            Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
+        }
     }
 }
 
 
 Describe  "Set Static Host List" {
     BeforeEach {
-        Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05 -host_entries_description "Add via PowerArubaCP"
+        if ($VersionBefore680 -eq 0) {
+            Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05 -host_entries_description "Add via PowerArubaCP"
+        }
     }
 
-    It "Set Static Host List (Name and Description)" {
+    It "Set Static Host List (Name and Description)" -Skip:$VersionBefore680 {
         Get-ArubaCPStaticHostList -name pester_SHL | Set-ArubaCPStaticHostList -name pester_SHL2 -description "Change via PowerArubaCP"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL2
         $shl.name | Should be "pester_SHL2"
         $shl.description | Should be "Change via PowerArubaCP"
     }
 
-    It "Set Static Host List (host_entries)" {
+    It "Set Static Host List (host_entries)" -Skip:$VersionBefore680 {
         $host_entries = @()
         $host_entries += @{ host_address = "00:01:02:03:04:06"; host_address_desc = "Change via PowerArubaCP" }
         Get-ArubaCPStaticHostList -name pester_SHL | Set-ArubaCPStaticHostList -host_entries $host_entries
@@ -207,13 +219,15 @@ Describe  "Set Static Host List" {
     }
 
     AfterEach {
-        Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
-        Get-ArubaCPStaticHostList -name pester_SHL2 | Remove-ArubaCPStaticHostList -noconfirm
+        if ($VersionBefore680 -eq 0) {
+            Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
+            Get-ArubaCPStaticHostList -name pester_SHL2 | Remove-ArubaCPStaticHostList -noconfirm
+        }
     }
 }
 Describe  "Remove Static Host List Member" {
 
-    It "Remove an entry of Static Host List with format list and type IPAddress " {
+    It "Remove an entry of Static Host List with format list and type IPAddress " -Skip:$VersionBefore680 {
         $shl = Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1, 192.0.2.2
         ($shl.host_entries).count | Should be "2"
         #Remove a entry...
@@ -223,7 +237,7 @@ Describe  "Remove Static Host List Member" {
         $shl.host_entries[0].host_address_desc | Should be "Add via PowerArubaCP"
     }
 
-    It "Remove an entry of Static Host List with format list and type MACAddress" {
+    It "Remove an entry of Static Host List with format list and type MACAddress" -Skip:$VersionBefore680 {
         $shl = Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05, 00:01:02:03:04:06
         ($shl.host_entries).count | Should be "2"
         #Remove a entry...
@@ -233,7 +247,7 @@ Describe  "Remove Static Host List Member" {
         $shl.host_entries[0].host_address_desc | Should be "Add via PowerArubaCP"
     }
 
-    It "Remove two entries of Static Host List with format list and type IPAddress " {
+    It "Remove two entries of Static Host List with format list and type IPAddress " -Skip:$VersionBefore680 {
         $shl = Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1, 192.0.2.2, 192.0.2.3
         ($shl.host_entries).count | Should be "3"
         #Remove 2 entries...
@@ -243,7 +257,7 @@ Describe  "Remove Static Host List Member" {
         $shl.host_entries[0].host_address_desc | Should be "Add via PowerArubaCP"
     }
 
-    It "Remove two entries of Static Host List with format list and type MACAddress" {
+    It "Remove two entries of Static Host List with format list and type MACAddress" -Skip:$VersionBefore680 {
         $shl = Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05, 00:01:02:03:04:06, 00:01:02:03:04:07
         ($shl.host_entries).count | Should be "3"
         #Remove 2 entries...
@@ -253,18 +267,20 @@ Describe  "Remove Static Host List Member" {
         $shl.host_entries[0].host_address_desc | Should be "Add via PowerArubaCP"
     }
 
-    It "Throw when remove all Entry" {
+    It "Throw when remove all Entry" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type MACAddress -host_entries_address 00:01:02:03:04:05
         { Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostListMember -host_entries_address 00:01:02:03:04:05 } | Should throw "You can't remove all entries. Use Remove-ArubaCPStaticHostList to remove Static Host List"
     }
 
     AfterEach {
-        Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
+        if ($VersionBefore680 -eq 0) {
+            Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
+        }
     }
 }
 Describe  "Remove Static Host List" {
 
-    It "Remove Static Host List by id" {
+    It "Remove Static Host List by id" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1 -host_entries_description "Add via PowerArubaCP"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
         $shl.name | Should be "pester_SHL"
@@ -275,7 +291,7 @@ Describe  "Remove Static Host List" {
         @($shl).count | should be 0
     }
 
-    It "Remove Static Host List by name (and pipeline)" {
+    It "Remove Static Host List by name (and pipeline)" -Skip:$VersionBefore680 {
         Add-ArubaCPStaticHostList -name pester_SHL -host_format list -host_type IPAddress -host_entries_address 192.0.2.1 -host_entries_description "Add via PowerArubaCP"
         $shl = Get-ArubaCPStaticHostList -name pester_SHL
         $shl.name | Should be "pester_SHL"
@@ -287,7 +303,10 @@ Describe  "Remove Static Host List" {
     }
 
     AfterEach {
-        Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
+        if ($VersionBefore680 -eq 0) {
+            Get-ArubaCPStaticHostList -name pester_SHL | Remove-ArubaCPStaticHostList -noconfirm
+        }
+
     }
 }
